+++ trimmerDecorator.rb
@@ -1,0 +1,5 @@
+class TrimmerDecorator < Decorator
+  def correct_name
+    @nameable.correct_name[0..9] if @nameable.correct_name.length > 10 || @nameable.correct_name
+  end
+end
